Update README.md
Project: Application testing https://vue-demo.daniel-avellaneda.com/
Assignment:
Create a repository on github called itbootcamp_final_project. A project must have commits after each completed unit, not just at the end.

Set up the project to add a valid version of chromedriver to it, and also add the testNG and selenium libraries to the pom.xml.
When creating a project, it is necessary to clearly separate packages for pages and tests. Example:
pages - where you store all classes that are pages
tests - where you store all classes that are tests

It is recommended that all code and commit messages be in English.

Mandatory use of wait-er that will wait for the opening of dialogs or similar loadings.

The base class that each test inherits from. This class should have:
baseUrl - url of the page https://vue-demo.daniel-avellaneda.com
beforeClass - where the driver with implicit wait and wait for page load is set and all the pages needed for testing are created
afterClass - where the page closes
beforeMethod - where the baseUrl page is loaded before each test

Only mandatory assertions are listed, add additional ones if necessary.

Divide the tests by classes:
Login Tests

Test #1: Visits the login page
assert:
Verify that the route /login appears in the url of the page


Test #2: Checks input types
assert:
Verify that the email input field for the type attribute has the value email
Verify that the password input field for the type attribute has the value password


Test #3: Displays errors when user does not exist
Data: random email and password using faker library
assert:
Verify that the error contains the message User does not exist
Verify that the /login route appears in the url of the page

Test #4: Displays errors when password is wrong
Data: email: admin@admin.com and arbitrary password

assert:
Verify that the error contains the message Wrong password
Verify that the /login route appears in the url of the page

Test #5: Login
The data:
email: admin@admin.com
password: 12345
assert:
Verify that the /home route appears in the url of the page

Test #6: Logout
assert:
Verify that the logout button is visible on the page
Verify that the /login route appears in the url of the page
Verify that after trying to open the /home route, the /login route appears in the url of the page (open with driver.get home page and check if it redirects you to login)

Signup Tests

Test #1: Visits the signup page
assert:
Verify that the /signup route appears in the url of the page

Test #2: Checks input types
assert:
Verify that the email input field for the type attribute has the value email
Verify that the password input field for the type attribute has the value password
Verify that the confirmation password input field for the type attribute has the value password

Test #3: Displays errors when user already exists
The data:
name: Test Test
email: admin@admin.com
password: 123654
confirm password: 123654
assert:
Verify that the error contains the message E-mail already exists
Verify that the /signup route appears in the url of the page

Test #4: Signup
The data:
name: Name and surname of the participant
email template: faker library - email
password: 123456
confirm password: 123456
assert:
Verify that the notification dialog contains the text IMPORTANT: Verify your account

Admin Cities Tests

Test #1: Visits the admin cities page and list cities
The data:
admin email: admin@admin.com
admin password: 12345
assert:
Verify that the /admin/cities route appears in the url of the page
Verify the existence of the logout button


Test #2: Create a new city
Data: random grad using faker library
assert:
Verify that the message contains the text Saved successfully

Test #3: Edit city
Data: the city created in test 2 with the same name is edited + - edited (example: Belgrade - Belgrade edited)
assert:
Verify that the message contains the text Saved successfully


Test #4: Search city
Data: edited city from test #3
assert:
Verify that there is text from the search in the Name column of the first row



Test #5: Delete city
Data: edited city from test #3
assert:
Enter the old name of the city in the search field
Wait until the number of rows in the table is 1
Verify that there is text from the search in the Name column of the first row
Click on the Delete button from the first row
Wait for the deletion dialog to appear
Click the Delete button from the dialog
Wait until the message display pop-up is visible
Verify that the message contains the text Deleted successfully
Auth Routes Tests

Test #1: Forbids visits to home url if not authenticated
assert:
Load the /home page when the user is not logged in
Verify that the route /login appears in the url of the page

Test #2: Forbids visits to profile url if not authenticated
assert:
Load /profile page
Verify that the route /login appears in the url of the page

Test #3: Forbids visits to admin cities url if not authenticated
assert:
Load the /admin/cities page
Verify that the route /login appears in the url of the page

Test #4: Forbids visits to admin users url if not authenticated
assert:
Load the /admin/users page
Verify that the route /login appears in the url of the page

Local Tests

Test #1: Set locale to ES
assert:
Set language to ES
Verify that the text Página de aterrizaje appears in the header of the page

Test #2: Set locale to EN
assert:
Set the language to EN
Verify that the Landing text appears in the header of the page

Test #3: Set locale to FR
assert:
Set language to FR
Verify that the text Page d'atterrissage appears in the header of the page



Profile Tests
Test #1: Edits profile
Data: random data using the faker library
assert:
Verify that the message Profile saved successfully is displayed
Verify that each input now for the value attribute has the value entered within the form